// src/sdk/generated/models/UserUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

/**
 * Schema for updating a user
 */
export type UserUpdate = {
  first_name?: string | null;
  last_name?: string | null;
  middle_name?: string | null;
  phone_number?: string | null;
  date_of_birth?: string | null;
  profile_picture_url?: string | null;
  bio?: string | null;
  language?: string | null;
  timezone?: string | null;
};